add swal

+++ angular-project/src/app/selecteddpo/dpo-detail/dpo-detail.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { DPOType } from 'src/app/shared/helpers/interface';
 import { DpoService } from 'src/app/shared/service/dpo.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-dpo-detail',
@@ -17,8 +18,30 @@ export class DpoDetailComponent implements OnInit {
   }
 
   onDeleteDPO(id: string): void {
-    this.dpoService.deletedpo(id);
-    this.router.navigate(['/home']);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this data!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.dpoService.deletedpo(id);
+        Swal.fire(
+          'Deleted!',
+          'Your data has been deleted.',
+          'success'
+        )
+        this.router.navigate(['/home']);
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your data is safe :)',
+          'error'
+        )
+      }
+    })
   }
 
 }
